Assorted changes ahead of C++ changes
The generate_sample code has never worked in this setup. I'm adding this in to have a place convenient to roll back to later, after trying to move the Kalman filter smoother over to C++, which may not work. This commit contains some changes to validation functions and unit tests which will be useful.